Integrate ZFramework game flow and Unity 6 tooling

+++ UnityProject/Packages/com.tengine.rts/Contracts/IWorldObject.cs
@@ -2,7 +2,7 @@ namespace TEngine.RTS
 {
     public interface IWorldObject
     {
-        int InstanceId { get; }
+        ulong InstanceId { get; }
         string Name { get; }
         bool TryGetCapability<T>(out T capability) where T : class;
     }
